Add parts photo
Most of the parts are in-house now. We have also finalized the power solution across the 3 sizes of the platform. Design work is back underway for the Amaptive wheelbase. That design will be carried into smaller Amaptive Micro and the Amaptive Senior. 

The wheelbase was chosen first since it will be possible to branch the project there initially. The wheelbase by itself will provide a modular RC platform which can be used in several ways.
Once the wheelbase is complete I will begin work on the brain module. 
This will ride on top of the wheelbase on a rail system. Allowing the brain module to crawl from one base to the next. This will get around being forced into charging. It will just transfer itself to a new wheelbase and continue on the mission while the previous base charges.